requirements.txt
absl-py==0.11.0
asgiref==3.7.2
astor==0.8.1
auth0-python==3.19.0
cached-property==1.5.2
cachetools==4.2.2
certifi>=2020.06.20
chardet==3.0.4
click==8.0.4
consulate==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-stripe==2.5.1
Django==3.2.23
django-admin-easy==0.6
django-admin-rangefilter==0.10.0
django-admin-relation-links==0.2.5
django-allauth==0.54.0
django-appconf==1.0.5
django-avatar==7.1.1
django-compat==1.0.15
django-embed-video==1.2.0
django-filter==23.2
django-guardian==2.4.0
django-hijack==3.1.4
django-imagekit==4.0.2
django-import-export==2.0.2
django-json-widget==0.2.0
django-mock-queries==2.1.5
django-nested-inline==0.4.2
django-ordered-model==3.4.1
django-safedelete==0.5.5
django-smuggler==1.0.1
django-sortedm2m==2.0.0
django-storages==1.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djrill==2.1.0
drf-api-tracking==1.8.4
drf-yasg==1.21.7
et-xmlfile==1.0.1
gast==0.2.2
google-api-core==2.11.1
google-cloud-storage==2.10.0
google-auth==2.22.0
google-cloud-vision==3.4.4
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.56.2
h5py==3.7.0
idna==2.8
imageio==2.9.0
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==3.1.2
joblib>=1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.6.1
mock==3.0.5
model-bakery==1.0.2
networkx==2.5
numpy>=1.7
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.5.2
pathos==0.3.0
Pillow>=5.4.1
protobuf==4.24.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2022.5
PyJWT==2.8.0
PyWavelets==1.4.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.0
rest_condition==1.0.3
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.10.1
setuptools-declarative-requirements==1.1.0
shortuuid==1.0.1
six==1.16.0
sorcery==0.2.2
sqlparse==0.4.4
tablib==2.0.0
termcolor==1.1.0
tqdm==4.64.1
uritemplate==3.0.1
urllib3==1.26.18
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
